Factoring out transient behavior for MOS
