+++ SpiceSharp/Components/Semiconductors/MOS/Level1/FrequencyBehavior.cs
@@ -183,7 +183,7 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level1
             var vgd = vgs - vds;
 
             var effectiveLength = _bp.Length - 2 * _mbp.LateralDiffusion;
-            var oxideCap = _modeltemp.OxideCapFactor * effectiveLength * _bp.Width;
+            var oxideCap = _mbp.OxideCapFactor * effectiveLength * _bp.Width;
 
             /* 
              * now we do the hard part of the bulk - drain and bulk - source
