+++ SpiceSharp/Components/Semiconductors/MOS/Level1/Mosfet1.cs
@@ -33,7 +33,7 @@ namespace SpiceSharp.Components
             // Add factories
             Behaviors.Add(typeof(TemperatureBehavior), () => new TemperatureBehavior(Name));
             Behaviors.Add(typeof(LoadBehavior), () => new LoadBehavior(Name));
-            Behaviors.Add(typeof(TransientBehavior), () => new TransientBehavior(Name));
+            Behaviors.Add(typeof(MosfetBehaviors.Common.TransientBehavior), () => new MosfetBehaviors.Common.TransientBehavior(Name));
             Behaviors.Add(typeof(FrequencyBehavior), () => new FrequencyBehavior(Name));
             Behaviors.Add(typeof(NoiseBehavior), () => new NoiseBehavior(Name));
         }
